Makefile.PL.
I can never get all of dzil's expansive deps installed.  It's failing
on Debian 8.8 and OpenBSD 6.1 in different places right now.

I think the only way people succeed there is by installing and accumulating
working versions of its deps over months or years, either from hacking on
other things or following dzil for a long time.  Trying to walk in to
dzil cold just fails every time.

The FatPacker packed wgd is failing to find bundled modules on Debian so my
Plan B for the installer is to download the zip from GitHub and build the
module, but dzil derailed that, so, Makefile.PL in the fork.

Please, if you Dist::Zilla, just commit the Makefile.PL it generates to the
repo.