fatpack this thing.
fatpack.trace is the edited output of running fatpack trace myscript.pl.
because of loading command modules on demand, fatpack didn't find
most of the WGDev modules.
also deleted all of the modules it found that are part of the perl
standard library listed in perldoc perlmodlib for 5.24.3.

to re-fat-pack after making changes:

* cpanm App::FatPacker
* update fatpack.trace with any new/removed modules
* fatpack packlists-for `cat fatpacker.trace` >packlists
* fatpack tree `cat packlists`
* fatpack file sbin/wgd > wgd
* commit and push changes including ./wgd